Added --asm-includes-location command line option that outputs the source file location that is represented by the generated code.
This allows tools like Compiler Explorer to map backwards from asm to source. It can be extended to debug output in future as it has no impact on the generated code (though has an impact on the compiler's memory usage.

Each ICode is now accompanied by the source file location that created it. This is passed through to the code generator which stores it in a parallel array to the generated code.

As the assembly is generated, it steps through both arrays and if the source file or source line number has changed, outputs a comment.